# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
...
	"info": {
		"title": "mini-project-2-ecomeernenrenren-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user": {
			"post": {
				"operationId": "CreateAccount",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"User name and password are too short. 4 < userName < 15 , 6 < password < 30",
								"User name is too short. 4 < userName < 15 ",
								"Password is too short. 6 < password < 30",
								"User name and password are too Long. 4 < userName < 15 , 6 < password < 30",
								"User name is too long. 4 < userName < 15",
								"Password is too long. 6 < password < 30",
								"User name is long password is short. 4 < userName < 15 , 6 < password < 30",
								"User name is short and password is long. 4 < userName < 15 , 6 < password < 30",
								"User created"
							],
							"x-nullable": false
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"userName": {
									"type": "string"
								},
								"password": {
									"type": "string"
								}
							},
							"title": "CreateAccountBody",
							"type": "object",
							"required": [
								"userName",
								"password"
							]
						}
					}
				]
			}
		},
		"/user/setAsAdmin/{_id}": {
			"put": {
				"operationId": "SetAsAdmin",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"No user found with this ID",
								"User is already Admin",
								"User is now Admin"
							],
							"x-nullable": false
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/user/removeAdmin/{_id}": {
			"put": {
				"operationId": "RemoveAdmin",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"No user found with this ID",
								"User is now Demoted",
								"User is not an admin anyway"
							],
							"x-nullable": false
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/user/subscribe/{_id}": {
			"put": {
				"operationId": "Subscribe",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"User is already premium",
								"User is now Premium"
							],
							"x-nullable": false
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/user/unsubscribe/{_id}": {
			"put": {
				"operationId": "Unsubscribe",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string",
							"enum": [
								"User now unsubscribed",
								"User is not a premium to unsubscribe"
							],
							"x-nullable": false
						}
					}
				},
				"security": [
					{
						"isLoggedIn": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "_id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "isLoggedIn",
			"name": "isLoggedIn",
			"in": "query"
		},
		"tsoa_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://swagger.io/api/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify things",
				"read:pets": "read things"
			}
		}
	}
}